Moved project specific interfaces and replace with generic
